backend/requirements.txt
bidict==0.21.2
certifi==2021.5.30
click==8.0.1
dnspython==1.16.0
eventlet==0.31.1
Flask==1.0.2
Flask-Login==0.4.1
Flask-Session==0.3.1
Flask-SocketIO==5.1.1
greenlet==1.1.1
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
numpy==1.21.2
python-engineio==4.2.1
python-socketio==5.4.0
six==1.11.0
Werkzeug==0.14.1
